Тестовое задание php-разработчик тг @chopin333
