+++ php/v2/Contractor.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace NW\WebService\References\Operations\Notification;
+
+/**
+ * @property Seller $Seller
+ */
+class Contractor
+{
+    const TYPE_CUSTOMER = 0;
+    public $id;
+    public $type;
+    public $name;
+
+    public static function getById(int $resellerId): ?self
+    {
+        //Не сталкивался с такими конструкциями) Создание себя же из статичного метода?
+        return new self($resellerId); // fakes the getById method
+    }
+
+    public function getFullName(): string
+    {
+        return $this->name . ' ' . $this->id;
+    }
+}
