+++ php/v2/NotificationEvents.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace NW\WebService\References\Operations\Notification;
+
+//Возможно не стоит создавать класс,а вынеси в константы
+class NotificationEvents
+{
+    const CHANGE_RETURN_STATUS = 'changeReturnStatus';
+    const NEW_RETURN_STATUS = 'newReturnStatus';
+}
