+++ php/v2/ReferencesOperations.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace NW\WebService\References\Operations\Notification;
+
+abstract class ReferencesOperation
+{
+    abstract public function doOperation(): array;
+
+    public function getRequest($pName): mixed
+    {
+        return $_REQUEST[$pName];
+    }
+}
