+++ php/v2/Seller.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace NW\WebService\References\Operations\Notification;
+
+use NW\WebService\References\Operations\Notification\Contractor;
+
+class Seller extends Contractor
+{
+}
