+++ php/v2/func.php
@@ -1,0 +1,12 @@
+<?php
+
+function getResellerEmailFrom()
+{
+    return 'contractor@example.com';
+}
+
+function getEmailsByPermit($resellerId, $event)
+{
+    // fakes the method
+    return ['someemeil@example.com', 'someemeil2@example.com'];
+}
