--- php/v2/others.php
@@ -1,75 +1,0 @@
-<?php
-
-namespace NW\WebService\References\Operations\Notification;
-
-/**
- * @property Seller $Seller
- */
-class Contractor
-{
-    const TYPE_CUSTOMER = 0;
-    public $id;
-    public $type;
-    public $name;
-
-    public static function getById(int $resellerId): self
-    {
-        return new self($resellerId); // fakes the getById method
-    }
-
-    public function getFullName(): string
-    {
-        return $this->name . ' ' . $this->id;
-    }
-}
-
-class Seller extends Contractor
-{
-}
-
-class Employee extends Contractor
-{
-}
-
-class Status
-{
-    public $id, $name;
-
-    public static function getName(int $id): string
-    {
-        $a = [
-            0 => 'Completed',
-            1 => 'Pending',
-            2 => 'Rejected',
-        ];
-
-        return $a[$id];
-    }
-}
-
-abstract class ReferencesOperation
-{
-    abstract public function doOperation(): array;
-
-    public function getRequest($pName)
-    {
-        return $_REQUEST[$pName];
-    }
-}
-
-function getResellerEmailFrom()
-{
-    return 'contractor@example.com';
-}
-
-function getEmailsByPermit($resellerId, $event)
-{
-    // fakes the method
-    return ['someemeil@example.com', 'someemeil2@example.com'];
-}
-
-class NotificationEvents
-{
-    const CHANGE_RETURN_STATUS = 'changeReturnStatus';
-    const NEW_RETURN_STATUS    = 'newReturnStatus';
-}
